active record: clear_all_connections on db connection failures
https://api.rubyonrails.org/classes/ActiveRecord/ConnectionAdapters/ConnectionHandler.html#method-i-clear_active_connections-21

> Returns any connections in use by the current thread back to the pool, and also returns connections to the pool cached by threads that are no longer alive.

Connection is retrieved explicitly here, so there should not be any "runaway"
connections.
`clear_active_connections!` makes sense to be done after every processing
cycle, but not (only) on errors.

But `clear_all_connections!` is suitable to be done on database
connection failures, as it explicitly empties the pool (disconnects
every pooled connection).

This should solve (or at least improve) issue when Fleiss is stuck in
exception loop after DB blips.